backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
rasterio==1.3.9
pillow==10.2.0
mercantile==1.2.1
numpy==1.26.4 